requirements.txt
snowflake-connector-python>=3.7.0
pandas>=2.0.0
python-dotenv>=1.0.0
dbt-core>=1.7.0
dbt-snowflake>=1.7.0
